samples/microsoft-agent-framework/multi-agent-servicebus/pyproject.toml
[project]
name = "maf-multi-agent-servicebus"
version = "0.1.0"
description = "Microsoft Agent Framework multi-agent pipeline with Azure Service Bus"
requires-python = ">=3.11"

dependencies = [
    "azure-ai-projects>=1.0.0b7",
    "azure-ai-agents>=1.0.0b4",
    "azure-identity>=1.15.0",
    "azure-servicebus>=7.12.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
shared = { workspace = true }
